перенос help на классы остается тесты
