+++ bot/dialogs/dialog_messages.yaml
@@ -171,7 +171,7 @@ bot:
           💎 Срок: {months} мес. ({price}₽)
           🕒 Ожидает подтверждения оплаты.
   errors:
-    unknown_command: "⚠️ Неизвестная команда. Пожалуйста, используйте кнопки ниже 👇 для выбора действия."
+    unknown_command: "⚠️ Неизвестная команда. Пожалуйста, используйте кнопки ниже/выше 👇👆 для выбора действия."
     unknown_command_admin: "⚠️ Неизвестная команда для админа. Пожалуйста, используйте кнопки ниже 👇 для выбора действия."
     help_limit_reached: |
                           ⚠️ Кажется, вы  {username} часто нажимаете кнопки.
